requirements.txt
Flask
SQLAlchemy
PyMySQL
PyYAML
